Rename FirebaseDB to Firebasefunctions & add login & register functions to FirebaseFunctions
